Update torch.full with dtype arg (#1934)
backport #1899

+++ src/garage/torch/_functions.py
@@ -68,7 +68,8 @@ def compute_advantages(discount, gae_lambda, max_path_length, baselines,
 
     """
     adv_filter = torch.full((1, 1, 1, max_path_length - 1),
-                            discount * gae_lambda)
+                            discount * gae_lambda,
+                            dtype=torch.float)
     adv_filter = torch.cumprod(F.pad(adv_filter, (1, 0), value=1), dim=-1)
 
     deltas = (rewards + discount * F.pad(baselines, (0, 1))[:, 1:] - baselines)
